Update ::Transformations datetime spec to make spec pass
Not sure if NS has changed their implementation since this test
was written. But currently, NS wants datetimes with its own UTC offset.
If an ActiveRecord model has an attribute with datetime
`12:34:56 -07:00`, and you want `12:34` to show up in the corresponding
field in NetSuite, then you need to send `12:34:56 -<NS_timezone_offset>`
to SuiteTalk.